feat: 宿主开关、2D 瓦片/精灵件与 Editor 素材工具（并同步服务端删除迁移帧）
- Core: 新增 EngineHostOptions + Game.ConfigureHost（后台运行 / 驱动器守卫 / AudioListener 归属），GridUtil、OrderedAsyncResult、Screenshot
- Presentation: 新增 TileWorld、SpriteFrameAnimator，扩展 UIWidgets / ObjectPool / Camera / Sound / FramePacing
- Resource: 新增 SpriteSet（批量异步预加载 → 按名同步取）
- Network: 移除 TCP/WS/QUIC 的 migrate 帧类型 3（服务端已删除该链路）
- Editor: 新增 PixelArtImport(+Settings/Postprocessor) 与 PanelPrefabBuilder
- Tests: 新增 GridUtil / TileWorld / SpriteFrameAnimator / SpriteSet / OrderedAsyncResult / EngineHostOptions 用例（EditMode）与表现池/音效用例（PlayMode）
- 索引 md 同步；本轮新增用例未在本机实跑

+++ Runtime/Presentation/FramePacing.cs
@@ -72,7 +72,11 @@ namespace CloverEngine
         /// </summary>
         public static float TryReadRefreshHz()
         {
-            try { return Screen.currentResolution.refreshRateRatio.value; }
+            // ⚠️ 2026-09-23 主 agent 修编译（12 个离线宿主**全部**红的唯一根因：它们都 Compile 本文件）：
+            //    `Screen.currentResolution.refreshRateRatio.value` 是 **double**（`RefreshRate.value`），
+            //    直接 return 到 `float` 签名会 **CS0266 隐式转换失败** ⇒ 整棵树编不过。
+            //    显式 `(float)` 修正，语义零改动（刷新率本身只有整数量级精度，float 完全够）。
+            try { return (float)Screen.currentResolution.refreshRateRatio.value; }
             catch (Exception) { return 0f; }
         }
